LayoutTests/imported/w3c/web-platform-tests/hr-time/raf-coarsened-time.https-expected.txt


FAIL raf-coarsened-time assert_approx_equals: document.timeline.currentTimeline should be the same as the rAF callback expected 95.94 +/- 0.01 but got 96
